Refactor all type classes that compile to Haskell function.
We are down from N type classes to 3, now using type families rather than
fundeps: CompileIO, Compile, and CompileLift. CompileIO compiles to a Haskell
function that returns its result in the IO monad, Compile compiles to a pure
function, and CompileLift compiles to a Template Haskell expression that is a
pure function.