Write CUDA function argument translators in CPS.
When we translate function arguments for a CUDA kernel call, we must allow the
translator to scope over the function call so that, e.g., we can call
```withForeignDevicePtr```. Therefore the ```toFunParams``` member of the
```ToFunParams``` class is now called with a continuation that takes the
translated function parameters as arguments.