Try a more reliable method for GC'ing GPU memory.
We keep all GPU memory in ForeignPtr's (ForeignDevicePtr is an alias for
ForeignPtr). However, we need to force unreferenced foreign pointers that hold
GPU memory to be finalized when a GPU memory allocation fails. I can't get that
to happen reliably, so instead we call performGC on every GPU memory allocation.