Clean up Lift, Unlift, and IsElem instances.
Lift and Unlift manage converting between Exp t a and types that are easier to
work with. For example, we can convert between Exp t (a, b) and (Exp t a, Exp t
b). These are moral isomorphisms. IsElem constrains which types can appear as
array elements in manifest arrays---delayed arrays can contain anything, of
course :)