No longer implicitly initialize a CUDA context.
Previously we tried to keep the user from having to explicitly initialize a CUDA
context. This required seq'ing a thunk (containing an unsafePerformIO) every
time we did something CUDA-related, which is disgusting. This was also not
thread safe since CUDA contexts are tied to a particular thread. Now the user
must take care of initializing a CUDA context.

As a consequence, we no longer compiled with the threaded runtime by default. In
particular, the test suite cannot be compiled with the threaded runtime because
it does not give us the opportunity to initialize a CUDA context on each test
runner thread.