Don't link twice with cuda libraries.
Since we depend on the cuda package, we are already being linked with libcuda
and libcudart. Linking in these libraries twice confuses cuda greatly, leading
to an "invalid context" error when creating an context for OpenGL interop. I
assume this has something to do with static initializers or thread-local state
or something.