Properly scope variables.
This patch reverts an earlier change that was introduced to mask a bug in the
way generated bindings were handled. Each let binding needs to introduce a new
delimited continuation for its body, otherwise bindings generated in the body,
which may refer to the bound variable, can escape the let. That wasn't
happening, although I'm not sure why and I'm not sure when the bug was fixed...