Add an array representation that serves to hint that computation is irregular.
The HintIrregular representation, whose type tag is I, hints that the work done
to compute the array it contains is irregular. Right now this causes parallel
for ('parfor') loops produced by loadP to be rewritten as irregular parallel for
('iparfor') loops. At the moment we don't have a compilation strategy for
irregular parallel for loops...that change is upcoming.